added artifact upload functionality
