+++ dapps/market-ui/src/components/cis2/Cis2BatchItemMetadataPrepare.tsx
@@ -1,266 +1,27 @@
-import { Buffer } from 'buffer/';
-import React, { FormEvent, useState } from 'react';
+import React, { useState } from 'react';
 
-import { CIS2, sha256 } from '@concordium/web-sdk';
-import {
-    Button, Card, CardActions, CardContent, CardMedia, Checkbox, FormControlLabel, FormGroup, Link,
-    Stack, SxProps, TextField, Typography
-} from '@mui/material';
+import { CIS2 } from '@concordium/web-sdk';
+import { SxProps } from '@mui/material';
 import { Theme } from '@mui/system';
 
-import { IPFS_GATEWAY_URL } from '../../Constants';
 import { Metadata } from '../../models/Cis2Client';
 import { PinataClient } from '../../models/PinataClient';
-import { tokenIdToNftImageFileName, tokenIdToNftMetadataFileName } from '../../models/Utils';
-import DisplayError from '../ui/DisplayError';
-import GetMintCardStep from './GetMintCardStep';
-import GetQuantityCardStep from './GetQuantityCardStep';
-import GetTokenIdCardStep from './GetTokenIdCardStep';
+import GetMintCardStep from './metadata-prepare-steps/GetMintCardStep';
+import GetQuantityCardStep from './metadata-prepare-steps/GetQuantityCardStep';
+import GetTokenIdCardStep from './metadata-prepare-steps/GetTokenIdCardStep';
+import UploadArtifactIpfsCardStep from './metadata-prepare-steps/UploadArtifactIpfsCardStep';
+import UploadImageIpfsCardStep from './metadata-prepare-steps/UploadImageIpfsCardStep';
+import UploadMetadataIpfsCardStep from './metadata-prepare-steps/UploadMetadataIpfsCardStep';
 
 const cardMediaSx: SxProps<Theme> = { maxHeight: "200px" };
 
 enum Steps {
-  GetTokenId,
-  UploadImage,
-  CreateMetadata,
-  GetQuantity,
-  Mint,
-}
-
-function UploadImageIpfsCardStep(props: {
-  tokenId: string;
-  file: File;
-  imageUrl: string;
-  pinata: PinataClient;
-  onDone: (input: { tokenId: string; imageIpfsUrl: string }) => void;
-}) {
-  const [state, setState] = useState({
-    tokenId: props.tokenId,
-    error: "",
-    imageUrl: props.imageUrl,
-    isUploadingImage: false,
-    imageIpfsUrl: "",
-  });
-
-  function submit() {
-    setState({ ...state, isUploadingImage: true });
-    props.pinata
-      .uploadFile(IPFS_GATEWAY_URL, props.file, tokenIdToNftImageFileName(props.file.name, props.tokenId))
-      .then((imageIpfsUrl) => {
-        setState({
-          ...state,
-          imageIpfsUrl: imageIpfsUrl,
-          isUploadingImage: false,
-          error: "",
-        });
-
-        props.onDone({ tokenId: props.tokenId, imageIpfsUrl });
-      })
-      .catch((error: Error) => setState({ ...state, error: error.message, isUploadingImage: false }));
-  }
-
-  return (
-    <Card variant="outlined">
-      <CardMedia component="img" image={props.imageUrl} alt="NFT" sx={cardMediaSx} />
-      <>
-        <CardContent>
-          <Typography>Upload File</Typography>
-          <Typography variant="caption">Token Id: {props.tokenId}</Typography>
-          <DisplayError error={state.error} />
-        </CardContent>
-        <CardActions>
-          <Button size="small" color="primary" onClick={submit} type="button" disabled={state.isUploadingImage}>
-            Upload Image
-          </Button>
-        </CardActions>
-      </>
-    </Card>
-  );
-}
-
-function UploadMetadataIpfsCardStep(props: {
-  tokenId: string;
-  imageUrl: string;
-  pinata: PinataClient;
-  imageIpfsUrl: string;
-  onDone: (data: { tokenId: string; metadataUrl: CIS2.MetadataUrl; metadata: Metadata }) => void;
-}) {
-  const [state, setState] = useState({
-    isUploadingMetadata: false,
-    metadataUrl: { url: "", hash: "" } as CIS2.MetadataUrl,
-    error: "",
-  });
-
-  const attributes = [
-    {
-      type: "string",
-      name: "Registry",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Serial Key",
-      value: "",
-    },
-    {
-      type: "number",
-      name: "Number of Credits",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Project Name",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Project Description",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Methodology",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Vintage",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Country",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Latitude",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Longitude",
-      value: "",
-    },
-    {
-      type: "string",
-      name: "Token Type",
-      value: "Register",
-    },
-  ];
-
-  function uploadMetadataClicked(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    const formData = new FormData(event.currentTarget);
-    const metadata: Metadata = {
-      name: formData.get("name")?.toString() || "",
-      description: formData.get("description")?.toString() || "",
-      display: {
-        url: props.imageIpfsUrl,
-      },
-      unique: !!formData.get("unique")?.toString(),
-      attributes: attributes
-        .map((a) => ({
-          type: a.type,
-          name: a.name,
-          value: formData.get(a.name)?.toString() || "",
-        }))
-        .filter((a) => !!a.value),
-    };
-    const includeHash = formData.get("includeHash")?.toString();
-    setState({ ...state, isUploadingMetadata: true });
-    props.pinata
-      .uploadJson(IPFS_GATEWAY_URL, metadata, tokenIdToNftMetadataFileName(props.tokenId))
-      .then((metadataIpfsUrl) => {
-        const metadataUrl = {
-          url: metadataIpfsUrl,
-          hash: includeHash ? sha256([Buffer.from(JSON.stringify(metadata))]).toString("hex") : "",
-        };
-        setState({
-          ...state,
-          metadataUrl,
-          isUploadingMetadata: false,
-          error: "",
-        });
-        props.onDone({ tokenId: props.tokenId, metadataUrl, metadata });
-      })
-      .catch((error: Error) => setState({ ...state, error: error.message, isUploadingMetadata: false }));
-  }
-
-  return (
-    <Card variant="outlined">
-      <CardMedia component="img" image={props.imageUrl} alt="NFT" sx={cardMediaSx} />
-      <>
-        <form onSubmit={(e) => uploadMetadataClicked(e)}>
-          <CardContent>
-            <Typography>Create Metadata</Typography>
-            <Typography variant="caption" component="div">
-              Token Id: {props.tokenId}
-            </Typography>
-            <Link href={props.imageIpfsUrl} variant="caption">
-              Image IPFS Url
-            </Link>
-            <Stack spacing={2}>
-              <FormGroup>
-                <TextField
-                  name="name"
-                  id="name"
-                  label="Name"
-                  variant="outlined"
-                  size="small"
-                  fullWidth={true}
-                  required={true}
-                  defaultValue={`Token ${props.tokenId}`}
-                />
-                <TextField
-                  multiline={true}
-                  name="description"
-                  id="description"
-                  label="Description"
-                  variant="outlined"
-                  size="small"
-                  fullWidth={true}
-                  required={true}
-                  defaultValue={`Image for token: ${props.tokenId}`}
-                />
-              </FormGroup>
-              <FormGroup>
-                <Stack spacing={1}>
-                  {attributes.map((a) => (
-                    <TextField
-                      key={a.name}
-                      name={a.name}
-                      id={a.name}
-                      label={a.name}
-                      variant="outlined"
-                      size="small"
-                      fullWidth={true}
-                      required={true}
-                      defaultValue={a.value}
-                      disabled={!!a.value}
-                    />
-                  ))}
-                </Stack>
-              </FormGroup>
-              <FormGroup>
-                <FormControlLabel
-                  control={<Checkbox defaultChecked name="includeHash" id="include-hash" />}
-                  label="Include Hash"
-                />
-                <FormControlLabel control={<Checkbox defaultChecked name="unique" id="unique" />} label="Unique" />
-              </FormGroup>
-            </Stack>
-            <DisplayError error={state.error} />
-          </CardContent>
-          <CardActions>
-            <Button size="small" color="primary" disabled={state.isUploadingMetadata} type="submit">
-              Create
-            </Button>
-          </CardActions>
-        </form>
-      </>
-    </Card>
-  );
+  GetTokenId = 0,
+  UploadImage = 1,
+  UploadArtifact = 2,
+  CreateMetadata = 3,
+  GetQuantity = 4,
+  Mint = 5,
 }
 
 function Cis2BatchItemMetadataPrepare(props: {
@@ -270,21 +31,33 @@ function Cis2BatchItemMetadataPrepare(props: {
   onDone: (data: { tokenId: string; tokenInfo: [CIS2.MetadataUrl, string] }) => void;
 }) {
   const pinata = new PinataClient(props.pinataJwtKey);
+  const [artifactUrl, setArtifactUrl] = useState("");
+  const [step, setStep] = useState(Steps.GetTokenId);
   const [state, setState] = useState({
     imageDisplayUrl: URL.createObjectURL(props.file),
-    step: Steps.GetTokenId,
     tokenId: props.tokenId,
     imageIpfsUrl: "",
     metadataUrl: { url: "", hash: "" } as CIS2.MetadataUrl,
     quantity: "",
   });
 
+  function goForward(skip = 1) {
+    setStep(step + skip);
+  }
+
   function tokenIdUpdated(tokenId: string) {
-    setState({ ...state, tokenId, step: Steps.UploadImage });
+    setState({ ...state, tokenId });
+    goForward();
   }
 
   function imageUploaded(tokenId: string, imageIpfsUrl: string) {
-    setState({ ...state, tokenId, step: Steps.CreateMetadata, imageIpfsUrl });
+    setState({ ...state, tokenId, imageIpfsUrl });
+    goForward();
+  }
+
+  function artifactUploaded(tokenId: string, ipfsUrl: string) {
+    setArtifactUrl(ipfsUrl);
+    goForward();
   }
 
   function metadataUploaded(tokenId: string, metadataUrl: CIS2.MetadataUrl, metadata: Metadata) {
@@ -292,20 +65,23 @@ function Cis2BatchItemMetadataPrepare(props: {
       // Since Unique is marked in metadata.
       // Quantity step can be skipped.
       const quantity = "1";
-      setState({ ...state, step: Steps.Mint, tokenId, quantity, metadataUrl });
+      setState({ ...state, tokenId, quantity, metadataUrl });
+      goForward(2);
       props.onDone({ tokenId, tokenInfo: [metadataUrl, quantity] });
     } else {
       // Let the user enter quantity.
-      setState({ ...state, tokenId, step: Steps.GetQuantity, metadataUrl });
+      setState({ ...state, tokenId, metadataUrl });
+      goForward();
     }
   }
 
   function quantityUpdated(tokenId: string, quantity: string) {
-    setState({ ...state, step: Steps.Mint, tokenId, quantity });
+    setState({ ...state, tokenId, quantity });
+    goForward();
     props.onDone({ tokenId, tokenInfo: [state.metadataUrl, quantity] });
   }
 
-  switch (state.step) {
+  switch (step) {
     case Steps.GetTokenId:
       return (
         <GetTokenIdCardStep
@@ -324,6 +100,18 @@ function Cis2BatchItemMetadataPrepare(props: {
           tokenId={state.tokenId}
           key={state.tokenId}
           onDone={(data) => imageUploaded(data.tokenId, data.imageIpfsUrl)}
+          sx={cardMediaSx}
+        />
+      );
+    case Steps.UploadArtifact:
+      return (
+        <UploadArtifactIpfsCardStep
+          pinata={pinata}
+          tokenId={state.tokenId}
+          key={state.tokenId}
+          onDone={({ tokenId, ipfsUrl }) => artifactUploaded(tokenId, ipfsUrl)}
+          onSkipped={() => goForward()}
+          sx={cardMediaSx}
         />
       );
     case Steps.CreateMetadata:
@@ -334,6 +122,7 @@ function Cis2BatchItemMetadataPrepare(props: {
           imageUrl={state.imageDisplayUrl}
           imageIpfsUrl={state.imageIpfsUrl}
           key={state.tokenId}
+          artifactIpfsUrl={artifactUrl}
           onDone={({ tokenId, metadataUrl, metadata }) => metadataUploaded(tokenId, metadataUrl, metadata)}
         />
       );
