+++ dapps/market-ui/src/components/cis2/metadata-prepare-steps/GetQuantityCardStep.tsx
@@ -4,7 +4,7 @@ import {
     Button, Card, CardActions, CardContent, CardMedia, SxProps, TextField, Theme, Typography
 } from '@mui/material';
 
-import DisplayError from '../ui/DisplayError';
+import DisplayError from '../../ui/DisplayError';
 
 const cardMediaSx: SxProps<Theme> = { maxHeight: "200px" };
